// server/types/customer/address.d.ts
import { RowDataPacket } from 'mysql2';

interface AddressRow extends RowDataPacket {
  id: number;
  company: string;
  email: string;
  phone: string;
  street_number: string;
  address1: string;
  address2: string;
  city: string;
  postal_code: string;
  country: string;
  customer_id: number;
}